% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{baby_naming_app}
\alias{baby_naming_app}
\title{Baby-naming APP (with reports and suggestions)}
\usage{
baby_naming_app(name, sex = 0, birth = NA)
}
\arguments{
\item{name}{Full name.}

\item{sex}{\code{0} = unknown, \code{-1} = female, \code{1} = male.}

\item{birth}{Birth year or \code{NA}.}
}
\value{
This function will output a list of results and "invisibly" return a final score (0-100).
}
\description{
Baby-naming APP (with reports and suggestions)
}
\examples{
testname=demodata[1, "name"]
sex=1  # 0 = unknown, -1 = female, 1 = male
baby_naming_app(testname, sex, 1995)
}
