% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\docType{package}
\name{ChineseNames-package}
\alias{ChineseNames-package}
\title{ChineseNames: Chinese Name Database 1930-2008}
\description{
A full database of Han Chinese names provided by
\emph{Beijing Meiming Science and Technology Company} and
originally obtained from the National Citizen Identity Information Center
(NCIIC) of China.
}
\references{
Please cite the following three references if you use this database.

Bao, H.-W.-S. (2020). ChineseNames: Chinese Name Database 1930-2008 [R package]. \link{https://github.com/psychbruce/ChineseNames}

Bao, H.-W.-S., Cai, H., DeWall, C. N., Gu, R., Chen, J., & Luo, Y. L. L. (2020). Name uniqueness predicts career choice and career achievement. Preprint at \emph{PsyArXiv} \link{https://doi.org/10.31234/osf.io/53j86}

Bao, H.-W.-S., Wang, J., & Cai, H. (2020). Blame crime on name? People with bad names are more likely to commit crime. Preprint at \emph{PsyArXiv} \link{https://doi.org/10.31234/osf.io/txhqg}
}
