% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{compute_name_index}
\alias{compute_name_index}
\title{Compute indexes of given names and surnames}
\usage{
compute_name_index(data = NULL, var.fullname = NULL,
  var.birthyear = NULL, index = c("NLen", "NU", "CCU", "NV", "NG",
  "SNU", "SNI"), return.namechar = FALSE)
}
\arguments{
\item{data}{\code{data.frame} or \code{data.table}.}

\item{var.fullname}{A variable (name) of full Chinese names.}

\item{var.birthyear}{[Optional] A variable (name) of birth year.}

\item{index}{[Optional] Which indexes to compute?

By default, it will compute all available variables.
\itemize{
  \item NLen: full-name length (2~4).
  \item NU: given-name uniqueness (1~6).
  \item CCU: character uniqueness in corpus (1~6).
  \item NV: given-name valence (1~5).
  \item NG: given-name gender (-1~1).
  \item SNU: surname uniqueness (1~6).
  \item SNI: surname initial (alphabetical order; 1~26).
}}

\item{return.namechar}{Whether to return name characters. Default is \code{FALSE}.}
}
\value{
A \code{data.frame} or \code{data.table} with a series of name indexes appended.
}
\description{
Compute indexes of given names and surnames
}
\examples{
demodata
compute_name_index(demodata, "name", "birth")  # adjust for birth cohort
compute_name_index(demodata, "name")  # no controlling for birth cohort
compute_name_index(demodata, "name", return.namechar=T)  # return name chars
}
