% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{familyname}
\alias{familyname}
\title{1,806 Chinese surnames and nationwide frequency.}
\format{
A data frame with 7 variables:
\describe{
  \item{\code{surname}}{surname (in Chinese)}
  \item{\code{compound}}{0 = single surname, 1 = compound surname}
  \item{\code{initial}}{initial letter (a-z)}
  \item{\code{initial.rank}}{initial order (1-26)}
  \item{\code{n.1930_2008}}{total counts in the database}
  \item{\code{ppm.1930_2008}}{proportion in population (ppm = parts per million)}
  \item{\code{surname.uniqueness}}{surname uniqueness}
}
}
\usage{
data(familyname)
}
\description{
1,806 Chinese surnames and nationwide frequency.
}
\details{
\url{https://psychbruce.github.io/ChineseNames}
}
