% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{population}
\alias{population}
\title{Population statistics for the Chinese name database.}
\usage{
data(population)
}
\description{
Population statistics for the Chinese name database.
}
\details{
\url{https://psychbruce.github.io/ChineseNames/}
}
