% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{top1000name.prov}
\alias{top1000name.prov}
\title{Top 1000 given names across 31 Chinese mainland provinces}
\usage{
top1000name.prov
}
\description{
Top 1000 given names across 31 Chinese mainland provinces
}
