% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\docType{package}
\name{ChineseNames-package}
\alias{ChineseNames-package}
\title{ChineseNames: Chinese Names Database 1930-2008}
\description{
A database of Han Chinese names provided by \href{https://www.qimingtong.com/}{Beijing Meiming Technology Company}.
}
\references{
To cite this package if you use the data in your research:

Bao, H.-W.-S. (2020). ChineseNames: Chinese Names Database 1930-2008 (R package). Retrieved from \link{https://github.com/psychbruce/ChineseNames}
}
