% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\docType{package}
\name{ChineseNames-package}
\alias{ChineseNames-package}
\title{ChineseNames: Chinese Names Database 1930-2008}
\description{
A database of Han Chinese names obtained from
the National Citizen Identity Information Center (NCIIC) of China
and provided by \emph{Beijing Meiming Science and Technology Company}.
}
\references{
If you use this database, please cite it as:

Bao, H.-W.-S. (2020). ChineseNames: Chinese Names Database 1930-2008 (R package). Retrieved from \link{https://github.com/psychbruce/ChineseNames}
}
