% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\docType{package}
\name{ChineseNames-package}
\alias{ChineseNames}
\alias{ChineseNames-package}
\title{ChineseNames: Chinese Name Database 1930-2008}
\source{
This database was provided by Beijing Meiming Science and Technology Company (in collaboration) and
originally obtained from the National Citizen Identity Information Center (NCIIC) of China in 2008.
}
\description{
A database of Chinese surnames and Chinese given names (1930-2008).
This database contains nationwide frequency statistics of
1,806 Chinese surnames and 2,614 Chinese characters used in given names,
covering about 1.2 billion Han Chinese population
(96.8\% of the Han Chinese household-registered population
born from 1930 to 2008 and still alive in 2008).
This package also contains a function for computing multiple features of
Chinese surnames and Chinese given names for scientific research (e.g.,
name uniqueness, name gender, name valence, and name warmth/competence).
}
\details{
Details are described in
\url{https://psychbruce.github.io/ChineseNames/}
}
\note{
This database does not contain any individual-level information
(so it does not leak personal privacy).
All data are at the name level or character level.
Extremely rare characters are not included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://psychbruce.github.io/ChineseNames/}
  \item Report bugs at \url{https://github.com/psychbruce/ChineseNames/issues}
}

}
\author{
\strong{Maintainer}: Han-Wu-Shuang Bao \email{baohws@foxmail.com} (\href{https://orcid.org/0000-0003-3043-710X}{ORCID})

}
