% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{compute_name_index}
\alias{compute_name_index}
\title{Easily compute variables of given names and surnames for scientific research}
\usage{
compute_name_index(
  data = NULL,
  var.fullname = NULL,
  var.birthyear = NULL,
  name = NA,
  birth = NA,
  index = c("NLen", "NU", "CCU", "NV", "NG", "SNU", "SNI"),
  return.namechar = TRUE,
  return.all = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} or \code{data.table}.}

\item{var.fullname}{Variable name of full Chinese names in your data (e.g., \code{"name"}).}

\item{var.birthyear}{[Optional] Variable name of birth year in your data (e.g., \code{"birth"}).}

\item{index}{[Optional] Which indexes to compute?

By default, it will compute all available variables.
\itemize{
  \item NLen: full-name length (2~4).
  \item NU: given-name uniqueness (1~6).
  \item CCU: character uniqueness in contemporary Chinese corpus (1~6).
  \item NV: given-name valence (1~5).
  \item NG: given-name gender (-1~1).
  \item SNU: surname uniqueness (1~6).
  \item SNI: surname initial (alphabetical order; 1~26).
}

For details about these variables, see \link{https://github.com/psychbruce/ChineseNames}}

\item{return.namechar}{Whether to return separate name characters. Default is \code{TRUE}.}

\item{return.all}{Whether to return all temporary variables in computing the final variables. Default is \code{FALSE}.}
}
\value{
A new \code{data.frame} or \code{data.table} with name variables appended.
}
\description{
To use this function, you also install the package \code{bruceR} from GitHub.
For an installation guide of \code{bruceR}, see \link{https://github.com/psychbruce/bruceR}
}
\examples{
## Compute for one name
myname=demodata[1, "name"]
mybirth=1995
compute_name_index(name=myname, birth=mybirth, index="NU")

## Compute for a dataset with a list of names
demodata  # a data frame
compute_name_index(demodata, "name", "birth")  # adjust for birth cohort
compute_name_index(demodata, "name")  # not adjust for birth cohort
compute_name_index(demodata, "name", return.all=T)  # return temporary variables
}
