% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{givenname}
\alias{givenname}
\title{Chinese given-name database (character level)}
\usage{
givenname
}
\description{
A list of 2,614 characters used in Chinese given names with their proportions in the Chinese population.
}
