% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{top1000name.prov}
\alias{top1000name.prov}
\title{Top 1,000 given names in 31 Chinese mainland provinces.}
\usage{
data(top1000name.prov)
}
\description{
Top 1,000 given names in 31 Chinese mainland provinces.
}
\details{
\url{https://github.com/psychbruce/ChineseNames}
}
