% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{top100name.year}
\alias{top100name.year}
\title{Top 100 given names (character combinations) for 6 birth cohorts}
\usage{
top100name.year
}
\description{
Top 100 given names (character combinations) for 6 birth cohorts
}
