% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{top100name.year}
\alias{top100name.year}
\title{Top 100 given names across 6 birth cohorts (pre-1960 to 2008)}
\usage{
top100name.year
}
\description{
Top 100 given names across 6 birth cohorts (pre-1960 to 2008)
}
