% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChineseNames.R
\name{top50char.year}
\alias{top50char.year}
\title{Top 50 given-name characters in 6 birth cohorts.}
\usage{
data(top50char.year)
}
\description{
Top 50 given-name characters in 6 birth cohorts.
}
\details{
\url{https://psychbruce.github.io/ChineseNames/}
}
